int? --> long?

+++ Octokit/Models/Request/NewRepository.cs
@@ -81,7 +81,7 @@ namespace Octokit
         /// <summary>
         /// Optional. Gets or sets the Id of the team to grant access to this repository. This is only valid when creating a repository for an organization.
         /// </summary>
-        public int? TeamId { get; set; }
+        public long? TeamId { get; set; }
 
         public bool? DeleteBranchOnMerge { get; set; }
