+++ Octokit/Models/Request/NewTeam.cs
@@ -60,7 +60,7 @@ namespace Octokit
         /// <summary>
         /// Id of a team to set as the parent team
         /// </summary>
-        public int? ParentTeamId { get; set; }
+        public long? ParentTeamId { get; set; }
 
         internal string DebuggerDisplay
         {
